requirements.txt
tensorflow
streamlit
pillow
numpy
plotly